+++ Source/EasyNetQ/Rpc/IRpcHeaderKeys.cs
@@ -1,0 +1,8 @@
+﻿namespace EasyNetQ.Rpc
+{
+    interface IRpcHeaderKeys
+    {
+        string IsFaultedKey { get; }
+        string ExceptionMessageKey { get; }
+    }
+}
