+++ Source/EasyNetQ/Rpc/RpcHeaderKeys.cs
@@ -1,0 +1,8 @@
+﻿namespace EasyNetQ.Rpc
+{
+    class RpcHeaderKeys : IRpcHeaderKeys
+    {
+        public string IsFaultedKey { get { return "IsFaulted"; } }
+        public string ExceptionMessageKey { get { return "ExceptionMessage"; } }
+    }
+}
